# generated/swagger/web/v1/web.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "web/v1/web.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Web"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/project/information": {
      "get": {
        "operationId": "Web_GetProjects",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetProjectsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Web"
        ]
      }
    },
    "/project/logs": {
      "get": {
        "operationId": "Web_Logs",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1Log"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1Log"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Web"
        ]
      }
    },
    "/project/logs/history": {
      "get": {
        "operationId": "Web_LogHistory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LogResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          }
        ],
        "tags": [
          "Web"
        ]
      }
    },
    "/project/{project}/inventory": {
      "get": {
        "operationId": "Web_GetProjectInventory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Project"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Web"
        ]
      }
    },
    "/project/{project}/service-dependency-graph": {
      "get": {
        "operationId": "Web_GetProjectServiceDependencyGraph",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GraphResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Web"
        ]
      }
    }
  },
  "definitions": {
    "LogKind": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "AGENT",
        "SERVICE"
      ],
      "default": "UNKNOWN"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1API": {
      "type": "object",
      "properties": {
        "tcp": {
          "$ref": "#/definitions/v1TcpAPI"
        },
        "rest": {
          "$ref": "#/definitions/v1RestAPI"
        },
        "grpc": {
          "$ref": "#/definitions/v1GrpcAPI"
        }
      }
    },
    "v1Application": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "project": {
          "type": "string"
        },
        "services": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Service"
          }
        }
      }
    },
    "v1ApplicationSnapshot": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "project": {
          "$ref": "#/definitions/v1ProjectSnapshot"
        }
      }
    },
    "v1Endpoint": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "visibility": {
          "type": "string"
        },
        "api": {
          "$ref": "#/definitions/v1API"
        },
        "application": {
          "type": "string"
        },
        "service": {
          "type": "string"
        },
        "namespace": {
          "type": "string"
        }
      }
    },
    "v1GetProjectsResponse": {
      "type": "object",
      "properties": {
        "projects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1GraphEdge": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        }
      }
    },
    "v1GraphNode": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "value": {
          "type": "boolean"
        }
      }
    },
    "v1GraphResponse": {
      "type": "object",
      "properties": {
        "nodes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GraphNode"
          }
        },
        "edges": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GraphEdge"
          }
        }
      }
    },
    "v1GrpcAPI": {
      "type": "object",
      "properties": {
        "proto": {
          "type": "string",
          "format": "byte"
        },
        "rpcs": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1RPC"
          }
        }
      }
    },
    "v1HTTPMethod": {
      "type": "string",
      "enum": [
        "GET",
        "POST",
        "PUT",
        "DELETE",
        "PATCH",
        "OPTIONS",
        "HEAD",
        "CONNECT",
        "TRACE"
      ],
      "default": "GET"
    },
    "v1Log": {
      "type": "object",
      "properties": {
        "at": {
          "type": "string",
          "format": "date-time"
        },
        "application": {
          "type": "string"
        },
        "service": {
          "type": "string"
        },
        "kind": {
          "$ref": "#/definitions/LogKind"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1LogResponse": {
      "type": "object",
      "properties": {
        "groups": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1LogSessionGroup"
          }
        }
      }
    },
    "v1LogSessionGroup": {
      "type": "object",
      "properties": {
        "session": {
          "$ref": "#/definitions/v1Session"
        },
        "logs": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Log"
          }
        }
      }
    },
    "v1Organization": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "domain": {
          "type": "string"
        }
      }
    },
    "v1PartialSnapshot": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "project": {
          "$ref": "#/definitions/v1ProjectSnapshot"
        },
        "applications": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1ApplicationSnapshot"
          }
        }
      }
    },
    "v1Project": {
      "type": "object",
      "properties": {
        "organization": {
          "$ref": "#/definitions/v1Organization"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "applications": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Application"
          }
        }
      }
    },
    "v1ProjectSnapshot": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "v1RPC": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "v1RestAPI": {
      "type": "object",
      "properties": {
        "openapi": {
          "type": "string",
          "format": "byte"
        },
        "routes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1RestRoute"
          }
        }
      }
    },
    "v1RestRoute": {
      "type": "object",
      "properties": {
        "methods": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1HTTPMethod"
          }
        },
        "path": {
          "type": "string"
        }
      }
    },
    "v1Service": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "application": {
          "type": "string"
        },
        "endpoints": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Endpoint"
          }
        }
      }
    },
    "v1Session": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "at": {
          "type": "string",
          "format": "date-time"
        },
        "partial": {
          "$ref": "#/definitions/v1PartialSnapshot"
        },
        "application": {
          "$ref": "#/definitions/v1ApplicationSnapshot"
        }
      }
    },
    "v1TcpAPI": {
      "type": "object"
    }
  }
}